Add special type for count number (if Qt developers add support for quint64 size type).

+++ src/hashfileinfostruct.h
@@ -7,7 +7,7 @@ typedef struct {
     QString fileName;
     QByteArray hash;
     bool checked;
-    uint groupID;
+    quint64 groupID;
     qint64 size;
     uint height, width;
 } HashFileInfoStruct;
